% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_to_df.R
\name{add_data_to_df}
\alias{add_data_to_df}
\title{Add data to a dataframe when querying and preparing}
\usage{
add_data_to_df(out = NULL, out_i = NULL, i, date_start_i, date_end_i)
}
\description{
Add data to a dataframe when querying and preparing
}
