% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_prepare_data.R
\name{get_and_prepare_data}
\alias{get_and_prepare_data}
\title{Queries and prepares monthly GHCND data for a selected location list, and between a start and end date}
\usage{
get_and_prepare_data(places_to_check, date_start, date_end)
}
\arguments{
\item{places_to_check}{string vector, date_start date in YYYY-MM-DD, date_end date in YYYY-MM-DD}
}
\value{
data frame
}
\description{
Queries and prepares monthly GHCND data for a selected location list, and between a start and end date
}
\examples{
places_to_check <- c("Hungary", "Spain", "United States")
date_start <- '2000-01-01'
date_end <- '2010-01-01'
get_and_prepare_data(places_to_check, date_start, date_end)
}
