% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_stations_list.R
\name{get_closest_stations_list}
\alias{get_closest_stations_list}
\title{Gets 10 closes station ids and takes the one with most of GHCND data.}
\usage{
get_closest_stations_list(geotags, date_start, date_end)
}
\arguments{
\item{geotags}{data frame, date_start date in YYYY-MM-DD, date_end date in YYYY-MM-DD}
}
\value{
list
}
\description{
Gets 10 closes station ids and takes the one with most of GHCND data.
}
