% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_place_for_stations.R
\name{get_place_for_stations}
\alias{get_place_for_stations}
\title{Gets closest station_id for latlon columns}
\usage{
get_place_for_stations(x, place_latlons)
}
\arguments{
\item{places_to_check}{string vector, date_start date in YYYY-MM-DD, date_end date in YYYY-MM-DD}
}
\value{
vector
}
\description{
Gets closest station_id for latlon columns
}
