% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnd_data_query.R
\name{ghcnd_data_query_for_stationidlist_between_dates}
\alias{ghcnd_data_query_for_stationidlist_between_dates}
\title{Queries GHCND data for a selected station id_list, and between a start and end date}
\usage{
ghcnd_data_query_for_stationidlist_between_dates(
  station_id_list,
  date_start,
  date_end
)
}
\arguments{
\item{station_id_list}{list, date_start date in YYYY-MM-DD, date_end date in YYYY-MM-DD}
}
\value{
data frame
}
\description{
Queries GHCND data for a selected station id_list, and between a start and end date
}
\examples{
station_id_list <- c("HU000012942", "LOE00105562", "NOE00112089", "SP000003195")
date_start <- '2000-01-01'
date_end <- '2010-01-01'
ghcnd_data_query_for_stationidlist_between_dates(station_id_list, date_start, date_end)
}
