% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_back.R
\name{nominatim_osm}
\alias{nominatim_osm}
\title{Getting geocodes based on openstreetmap public API}
\usage{
nominatim_osm(address = NULL)
}
\description{
Getting geocodes based on openstreetmap public API
}
